Change of plans: REST instead of gRPC
To make full use of gRPC, we'll need to redefine many of the types from
Warp. With REST, we can just serialize the existing types as JSON
payloads to REST calls and REST is simpler anyway.

We can always swich to gRPC later if REST turns out to be no good enough
for us.